backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-groq
langgraph
langgraph-checkpoint-sqlite
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
pydantic
